% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_cancel}
\alias{batch_cancel}
\title{Cancel batch job}
\usage{
batch_cancel(batch_id)
}
\arguments{
\item{batch_id}{(chr) the batch id to cancel}
}
\value{
(tibble) information about the cancelled batch
}
\description{
Cancel batch job
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_cancel("batch_abc123")

  batch_file_info <- batch_upload_file("abc123.jsonl")
  batch_job_info <- batch_file_info[["id"]] |>
    batch_create()
  batch_cancelled <- batch_job_info[["id"]] |>
    batch_cancel()
  batch_cancelled
}
}
\seealso{
Other batch: 
\code{\link{batch_create}()},
\code{\link{batch_list}()},
\code{\link{batch_result}()},
\code{\link{batch_status}()},
\code{\link{batch_upload_file}()}
}
\concept{batch}
