% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_create}
\alias{batch_create}
\title{Create batch}
\usage{
batch_create(input_file_id)
}
\arguments{
\item{input_file_id}{(chr) id of the input file}
}
\value{
(tibble) information about the created batch
}
\description{
Create batch
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_create("file-abc123")

  batch_file_info <- file_upload("abc123.jsonl")
  batch_job_info <- batch_file_info[["id"]] |>
    batch_create()
  batch_job_info
}
}
\seealso{
Other batches: 
\code{\link{batch_cancel}()},
\code{\link{batch_list}()},
\code{\link{batch_result}()},
\code{\link{batch_status}()},
\code{\link{file_upload}()}
}
\concept{batches}
