% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_list}
\alias{batch_list}
\title{List all batches}
\usage{
batch_list(n = 10)
}
\arguments{
\item{n}{(int) number of batches to retrieve}
}
\value{
(tibble) information about the batches
}
\description{
List all batches
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_list()
  batch_list(2)
  batch_list(Inf)
}
}
\seealso{
Other batches: 
\code{\link{batch_cancel}()},
\code{\link{batch_create}()},
\code{\link{batch_result}()},
\code{\link{batch_status}()},
\code{\link{file_upload}()}
}
\concept{batches}
