% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_result}
\alias{batch_result}
\title{Retrieve batch results}
\usage{
batch_result(batch_id, simplify = TRUE)
}
\arguments{
\item{batch_id}{(chr) the batch id to retrieve}

\item{simplify}{(lgl, default TRUE) whether to simplify the output,
i.e. return only the response body as for single standard
completions (the default), or the full response.}
}
\value{
(list) if simplify is TRUE, a list of the batch results;
otherwise, a list of the full responses.
}
\description{
Retrieve batch results
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_file_info <- file_upload("abc123.jsonl")
  batch_job_info <- batch_file_info[["id"]] |>
    batch_create()
  batch_status <- batch_job_info[["id"]] |>
    batch_status()

  # once the batch is completed
  if (batch_status[["status"]] == "completed") {
    results <- batch_status[["id"]] |>
      batch_result()
    res <- results |>
      purrr::map_chr(get_content)
    res
  }

  full_results <- batch_status[["id"]] |>
    batch_result(simplify = FALSE)
  str(full_results, 2)
  full_res <- full_results |>
    purrr::map_chr(\(x) get_content(x[["response"]][["body"]]))
  full_res

  identical(res, full_res)
}
}
\seealso{
Other batches: 
\code{\link{batch_cancel}()},
\code{\link{batch_create}()},
\code{\link{batch_list}()},
\code{\link{batch_status}()},
\code{\link{file_upload}()}
}
\concept{batches}
