% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_status}
\alias{batch_status}
\title{Retrieve batch status}
\usage{
batch_status(batch_id = "")
}
\arguments{
\item{batch_id}{(chr) the batch id to check}
}
\value{
(tibble) information about the batch status
}
\description{
This function retrieves the status of a batch.
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.

For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_status("batch_abc123")

  batch_file_info <- file_upload("abc123.jsonl")
  batch_job_info <- batch_file_info[["id"]] |>
    batch_create()
  batch_status <- batch_job_info[["id"]] |>
    batch_status()
  batch_status
}
}
\seealso{
Other batches: 
\code{\link{batch_cancel}()},
\code{\link{batch_create}()},
\code{\link{batch_list}()},
\code{\link{batch_result}()},
\code{\link{file_upload}()}
}
\concept{batches}
