% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_upload_file}
\alias{batch_upload_file}
\title{Upload batch file}
\usage{
batch_upload_file(jsonl_path)
}
\arguments{
\item{jsonl_path}{(chr) path to jsonl file to upload}
}
\value{
(tibble) information about the uploaded file
}
\description{
Upload batch file
}
\details{
For more information, see the
\href{https://platform.openai.com/docs/api-reference/batch}{OpenAI API documentation}.
}
\examples{
if (FALSE) {
  batch_file_info <- batch_upload_file("abc123.jsonl")
  batch_file_info
}
}
\seealso{
Other batch: 
\code{\link{batch_cancel}()},
\code{\link{batch_create}()},
\code{\link{batch_list}()},
\code{\link{batch_result}()},
\code{\link{batch_status}()}
}
\concept{batch}
