% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_prompt.R
\name{compose_prompt}
\alias{compose_prompt}
\alias{compose_prompt_system}
\alias{compose_prompt_user}
\title{Create a prompt to ChatGPT}
\usage{
compose_prompt(
  role = "",
  context = "",
  task = "",
  instructions = "",
  output = "",
  style = "",
  examples = "",
  text = "",
  delimiter = if (text == "") "" else "\\"\\"\\"\\""
)

compose_prompt_system(role = "", context = "")

compose_prompt_user(
  task = "",
  instructions = "",
  output = "",
  style = "",
  examples = "",
  text = "",
  delimiter = if (text == "") "" else "\\"\\"\\"\\""
)
}
\arguments{
\item{role}{(chr) The role that ChatGPT should play}

\item{context}{(chr) The context behind the task required}

\item{task}{(chr) The tasks ChatGPT should assess}

\item{instructions}{(chr) Description of steps ChatGPT should follow}

\item{output}{(chr) The type/kind of output required}

\item{style}{(chr) The style ChatGPT should use in the output}

\item{examples}{(chr) Some examples of correct output}

\item{text}{(chr) Additional text to embed in the prompt}

\item{delimiter}{(chr) delimiters for the \code{text} to embed, a sequence of
three identical symbols is suggested}
}
\value{
(chr) the glue of all the prompts components

(chr) The complete system prompt

(chr) The complete user prompt
}
\description{
Questa funzione è un semplice wrapper per comporre un buon prompt per
ChatGPT. L'output non è altro che la giustapposizione su righe separate delle
varie componenti (con il testo addizionale racchiuso tra i delimitatori in
fondo al prompt). Dunque il suo utilizzo è più che altro focalizzato è utile
per ricordare e prendere l'abitudine di inserire le componenti utili per un
buon prompt.
}
\section{Functions}{
\itemize{
\item \code{compose_prompt_system()}: 

\item \code{compose_prompt_user()}: 

}}
\examples{
if (FALSE) {
  compose_prompt(
    role = "Sei l'assistente di un docente universitario.",
    context = "
      Tu e lui state preparando un workshop sull'utilizzo di ChatGPT
      per biostatisitci ed epidemiologi.",
    task = "
      Il tuo compito è trovare cosa dire per spiegare cosa sia una
      chat di ChatGPT agli studenti, considerando che potrebbe
      esserci qualcuno che non ne ha mai sentito parlare (e segue
      il worksho incuriosito dal titolo o dagli amici).",
    output = "
      Riporta un potenziale dialogo tra il docente e gli studenti
      che assolva ed esemplifichi lo scopo descritto.",
   style = "Usa un tono amichevole, colloquiale, ma preciso."
 )
}
if (FALSE) {
  msg_sys <- compose_prompt_system(
    role = "Sei l'assistente di un docente universitario.",
    context = "
      Tu e lui state preparando un workshop sull'utilizzo di ChatGPT
      per biostatisitci ed epidemiologi."
 )
}
if (FALSE) {
  msg_usr <- compose_prompt_user(
    task = "
      Il tuo compito è trovare cosa dire per spiegare cosa sia una
      chat di ChatGPT agli studenti, considerando che potrebbe
      esserci qualcuno che non ne ha mai sentito parlare (e segue
      il worksho incuriosito dal titolo o dagli amici).",
    output = "
      Riporta un potenziale dialogo tra il docente e gli studenti
      che assolva ed esemplifichi lo scopo descritto.",
   style = "Usa un tono amichevole, colloquiale, ma preciso."
 )
}
}
