% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_prompt.R
\name{create_usr_data_prompter}
\alias{create_usr_data_prompter}
\title{Create a function to prompt the user for data}
\usage{
create_usr_data_prompter(
  task = "",
  instructions = "",
  output = "",
  style = "",
  examples = ""
)
}
\arguments{
\item{task}{(chr) The task ChatGPT should assess}

\item{instructions}{(chr) Description of steps ChatGPT should follow}

\item{output}{(chr) The type/kind of output required}

\item{style}{(chr) The style ChatGPT should use in the output}

\item{examples}{(chr) Some examples of correct output}
}
\value{
(function) a function that can be used to prompt the user,
accepting a string of text as input and returning the complete
prompt based on the desired context.
}
\description{
This function create a function that can be used to prompt the user
for data in a specific context. Given the interested context, the
function created will accept a string of text as input and return the
complete prompt based on the desired context.
}
\examples{
prompter <- create_usr_data_prompter(
  task = "Your task is to extract information from a text provided.",
  instructions = "
    You should extract the first and last words of the text.",
  output = "
    Return the first and last words of the text separated by a dash,
     i.e., `first - last`.",
  style = "
    Do not add any additional information, return only the requested
    information.",
  examples = "
    text: 'This is an example text.'
    output: 'This - text'
    text: 'Another example text!!!'
    output: 'Another - text'"
)
prompter("This is an example text.")
prompter("Another example text!!!")

# You can also use it with a data frame to programmaically create
# prompts for each row of a data frame's column.
db <- data.frame(
  text = c("This is an example text.", "Another example text!!!")
)
db$text |> purrr::map_chr(prompter)

}
