% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_list}
\alias{file_list}
\title{List uploaded file}
\usage{
file_list(purpose = NULL)
}
\arguments{
\item{purpose}{(chr) Only return files with the given purpose. Can be
omitted (i.e., NULL), "batch" or "batch_output".}
}
\value{
(tibble) information about the uploaded files
}
\description{
List uploaded file
}
\examples{
file_list()
file_list("batch")
file_list("batch_output")
}
\seealso{
Other files: 
\code{\link{file_delete}()},
\code{\link{file_info}()},
\code{\link{file_retrieve}()},
\code{\link{file_upload}()}
}
\concept{files}
