% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_retrieve}
\alias{file_retrieve}
\title{Retrieve file content}
\usage{
file_retrieve(file_id, convert_json = FALSE)
}
\arguments{
\item{file_id}{(chr) The id of the file to retrieve.}

\item{convert_json}{(lgl) If TRUE, convert the response to JSON.}
}
\value{
(chr) The content of the file.
}
\description{
Retrieve file content
}
\examples{
 sys_prompt <- compose_sys_prompt("You are a funny assistant.")
 usr_prompt <- compose_usr_prompt("Tell me a joke ending in:")
 prompter <- create_usr_data_prompter(usr_prompt = usr_prompt)
 text <-  c(
   "deadly boring!",
   "A bit boring, but interesting",
   "How nice, I loved it!"
 )

 jsonl_text <- text |>
   purrr::map(
     \(x) {
       compose_prompt_api(
         sys_prompt = sys_prompt,
         usr_prompt = prompter(x)
       )
     }
   ) |>
   create_jsonl_records()
 out_jsonl_path <- write_jsonl_files(jsonl_text, tempdir())

uploaded_file_info <- file_upload(out_jsonl_path)
file_content <- file_retrieve(uploaded_file_info[["id"]])
file_delete(uploaded_file_info[["id"]])
}
\seealso{
Other files: 
\code{\link{file_delete}()},
\code{\link{file_info}()},
\code{\link{file_list}()},
\code{\link{file_upload}()}
}
\concept{files}
