% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpteasyr-package.R
\docType{package}
\name{gpteasyr-package}
\alias{gpteasyr}
\alias{gpteasyr-package}
\title{gpteasyr: A basic and simple interface to OpenAI’s GPT API}
\description{
The goal of 'gpteasyr' is to provide a basic and simple interface to OpenAI's GPT API (and other compatible APIs). The package is also designed to work with (i.e., to query on) dataframes/tibbles, and to simplify the process of querying the API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/gpteasyr}
  \item \url{https://CorradoLanera.github.io/gpteasyr/}
  \item Report bugs at \url{https://github.com/CorradoLanera/gpteasyr/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

}
\keyword{internal}
