% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_py.R
\name{setup_py}
\alias{setup_py}
\title{Setup Python environment}
\usage{
setup_py(venv_name = "r-gpt-venv", ask = interactive())
}
\arguments{
\item{venv_name}{(chr, default "r-gpt-venv") The name of the virtual
environment to be created.}

\item{ask}{(lgl, default \code{TRUE} if in interactive session, \verb{FALSE otherwise}) If \code{TRUE}, the user is asked if they want to create the
virtual environment.}
}
\value{
(lgl) \code{TRUE} if the virtual environment was created, \code{FALSE}
otherwise.
}
\description{
This function creates a virtual environment and installs the \code{openai}
package in it.
}
\examples{
if (FALSE) {
  library(ubep.gpt)
  setup_py()

  prompt <- compose_prompt_api(
    sys_prompt = "You are the assistant of a university professor.",
    usr_prompt = "Tell me about the last course you provided."
  )

  res <- query_gpt(
      prompt = prompt,
      use_py = TRUE
    ) |>
      get_content()

   cat(res)
}
}
