% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubep.gpt-package.R
\docType{package}
\name{ubep.gpt-package}
\alias{ubep.gpt}
\alias{ubep.gpt-package}
\title{ubep.gpt: A basic/simple interface to OpenAI’s GPT API}
\description{
The goal of '{ubep.gpt}' is to provide a basic/simple interface to OpenAI's GPT API. The package is designed to work with (i.e., to query on) dataframes/tibbles, and to simplify the process of querying the API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/ubep.gpt}
  \item \url{https://ubesp-dctv.github.io/ubep.gpt/}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/ubep.gpt/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

}
\keyword{internal}
