% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_prompt.R
\name{compose_prompt}
\alias{compose_prompt}
\alias{compose_prompt_system}
\alias{compose_prompt_user}
\title{Create a prompt to ChatGPT}
\usage{
compose_prompt(
  role = "",
  context = "",
  task = "",
  instructions = "",
  output = "",
  style = "",
  examples = "",
  text = "",
  delimiter = if (text == "") "" else "\\"\\"\\"\\""
)

compose_prompt_system(role = "", context = "")

compose_prompt_user(
  task = "",
  instructions = "",
  output = "",
  style = "",
  examples = "",
  text = "",
  delimiter = if (text == "") "" else "\\"\\"\\"\\""
)
}
\arguments{
\item{role}{(chr) The role that ChatGPT should play}

\item{context}{(chr) The context behind the task required}

\item{task}{(chr) The tasks ChatGPT should assess}

\item{instructions}{(chr) Description of steps ChatGPT should follow}

\item{output}{(chr) The type/kind of output required}

\item{style}{(chr) The style ChatGPT should use in the output}

\item{examples}{(chr) Some examples of correct output}

\item{text}{(chr) Additional text to embed in the prompt}

\item{delimiter}{(chr) delimiters for the \code{text} to embed, a sequence
of three identical symbols is suggested}
}
\value{
(chr) the glue of all the prompts components

(chr) The complete system prompt

(chr) The complete user prompt
}
\description{
This function is a simple wrapper to compose a good prompt for
ChatGPT. The output is nothing more than the juxtaposition on
separate lines of the various components (with the additional text
enclosed between the delimiters at the bottom of the prompt). So its
use is more focused and useful for remembering and getting used to
entering the components useful for a good prompt.
}
\section{Functions}{
\itemize{
\item \code{compose_prompt_system()}: 

\item \code{compose_prompt_user()}: 

}}
\examples{
if (FALSE) {
  compose_prompt(
    role = "Sei l'assistente di un docente universitario.",
    context = "
      Tu e lui state preparando un workshop sull'utilizzo di ChatGPT
      per biostatisitci ed epidemiologi.",
    task = "
      Il tuo compito è trovare cosa dire per spiegare cosa sia una
      chat di ChatGPT agli studenti, considerando che potrebbe
      esserci qualcuno che non ne ha mai sentito parlare (e segue
      il worksho incuriosito dal titolo o dagli amici).",
    output = "
      Riporta un potenziale dialogo tra il docente e gli studenti
      che assolva ed esemplifichi lo scopo descritto.",
   style = "Usa un tono amichevole, colloquiale, ma preciso."
 )
}
if (FALSE) {
  msg_sys <- compose_prompt_system(
    role = "Sei l'assistente di un docente universitario.",
    context = "
      Tu e lui state preparando un workshop sull'utilizzo di ChatGPT
      per biostatisitci ed epidemiologi."
 )
}
if (FALSE) {
  msg_usr <- compose_prompt_user(
    task = "
      Il tuo compito è trovare cosa dire per spiegare cosa sia una
      chat di ChatGPT agli studenti, considerando che potrebbe
      esserci qualcuno che non ne ha mai sentito parlare (e segue
      il worksho incuriosito dal titolo o dagli amici).",
    output = "
      Riporta un potenziale dialogo tra il docente e gli studenti
      che assolva ed esemplifichi lo scopo descritto.",
   style = "Usa un tono amichevole, colloquiale, ma preciso."
 )
}
}
